Data Engineering Module/scripts/requirements.txt
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
requests
